PromptRange v0.4.7.2 修复最终得分的小数问题

+++ src/Extensions/Senparc.Xncf.PromptRange/Domain/Models/DatabaseModel/Dto/PromptResultDto.cs
@@ -61,7 +61,7 @@ namespace Senparc.Xncf.PromptRange.Models.DatabaseModel.Dto
         /// <summary>
         /// 最终得分
         /// </summary>
-        public int FinalScore { get; set; }
+        public decimal FinalScore { get; set; }
 
 
         /// <summary>
